Address reviews, move FetcherAndSampler to neighbor_sampler.py
